Fix for bug# JUDDI-1.  When searching by category bag for any of the four core objects (business, service, binding or tmodel) we were searching by keyName and keyValue which is incorrect.
In all cases only TModelKey and keyValue from the keyedRefernce are required and keyName is insignificant and can be ignored. There is one caveat however. If TModelKey is not specified then TModelKey for "uddi-org:general_keywords" should be used.

See http://nagoya.apache.org/jira/browse/JUDDI-1 for more details.


git-svn-id: https://svn.apache.org/repos/asf/webservices/juddi/trunk@262893 13f79535-47bb-0310-9956-ffa450edef68