Updated to set "" tModelKey in CategoryBag KeyedReferences when TModelKey as been specified per UDDI v2.04 Prog. API Section 4.3.5.1:
Normally, the tModelKey MUST contain a valid tModelKey for the
keyedReference to be valid. However, in the case of a keyedReference that is used in a categoryBag, the
tModelKey may be omitted or specified as a zero-length string to indicate that the taxonomy being used is
uddi-org:general_keywords. When it is omitted in this manner, the UDDI registry will insert the proper key
during the save_xx operation.


git-svn-id: https://svn.apache.org/repos/asf/webservices/juddi/trunk@262898 13f79535-47bb-0310-9956-ffa450edef68